feat(typings): moderation app typings
